Creating common place for reused tests and testing models
